requirements.txt
# Dataset for Translator - Requirements
# Python 3.11

# =============================================================================
# Phase 1: PDF Extraction
# =============================================================================
docling>=2.15.0
docling-core>=2.0.0

# =============================================================================
# Phase 2: Sentence Chunking
# =============================================================================
# Built-in regex is sufficient, no extra deps needed

# =============================================================================
# Phase 3: Embedding & Alignment
# =============================================================================
sentence-transformers>=3.3.0
faiss-gpu>=1.7.2  # Use faiss-cpu if no GPU available
torch>=2.0.0
transformers>=4.45.0

# =============================================================================
# Phase 4: CPO Training
# =============================================================================
trl>=0.12.0
peft>=0.13.0
accelerate>=1.0.0
bitsandbytes>=0.44.0
datasets>=3.0.0

# =============================================================================
# Utilities
# =============================================================================
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
jsonlines>=4.0.0
wandb>=0.18.0  # For experiment tracking

# =============================================================================
# Evaluation
# =============================================================================
sacrebleu>=2.4.0
comet-ml>=3.35.0
unbabel-comet>=2.2.0  # COMET metric for translation quality
